expand seButtton to include icon
