+++ src/editor/components/index.js
@@ -12,4 +12,3 @@ import './seListItem.js'
 import './seColorPicker.js'
 import './seSelect.js'
 import './seText.js'
-import './tsButton.js'
