+++ src/editor/extensions/ext-layer_view/ext-layer_view.js
@@ -22,7 +22,7 @@ const loadExtensionTranslation = async function (svgEditor) {
 
 export default {
   name,
-  async init (_S) {
+  async init(_S) {
     const svgEditor = this
     const { svgCanvas } = svgEditor
     const { $id, $click } = svgCanvas
@@ -57,7 +57,7 @@ export default {
     return {
       name: svgEditor.i18next.t(`${name}:name`),
       // The callback should be used to load the DOM with the appropriate UI items
-      layersChanged () {
+      layersChanged() {
         if ($id('tool_layerView').pressed) {
           updateLayerView()
         } if (svgEditor.configObj.curConfig.layerView) {
@@ -66,17 +66,17 @@ export default {
           updateLayerView()
         }
       },
-      layerVisChanged () {
+      layerVisChanged() {
         if ($id('tool_layerView').pressed) {
           $id('tool_layerView').pressed = !$id('tool_layerView').pressed
         }
       },
-      callback () {
+      callback() {
         const buttonTemplate = document.createElement('template')
         const title = `${name}:buttons.0.title`
         const key = `${name}:buttons.0.key`
         buttonTemplate.innerHTML = `
-      <ts-button id="tool_layerView" title="${title}" shortcut="${key}" icon="bi bi-stack"></ts-button>`
+        <se-button id="tool_layerView" title="${title}" shortcut="${key}" icon="bi bi-stack"></se-button>`
         $id('editor_panel').append(buttonTemplate.content.cloneNode(true))
         $click($id('tool_layerView'), clickLayerView.bind(this))
       }
